pyproject.toml
[project]
name = "marketplace-api"
version = "0.1.0"
description = ""
authors = [
    {name = "yiwywo18ypwtp1",email = "negalapetrovna77@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "pydantic-settings (>=2.14.1,<3.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "websockets (>=16.0,<17.0)",
    "celery (>=5.6.3,<6.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "psycopg2-binary (>=2.9.12,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "greenlet (>=3.5.0,<4.0.0)",
    "pydantic[email] (>=2.13.4,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
